abstracted out shader source into Shader class
Whenever we create a new shader instance the respective glsl file's path will be provided to it and by using regex we're processing the code inside that file to abstract out the vertex and fragment source code. The compile and linking logic is moved from LevelEditorScene to Shader's compile() method. use() and detach() methods also added to Shader class.